unet_api/requirements.txt
# Web框架核心
flask==3.0.3
werkzeug==3.0.6
itsdangerous==2.2.0  # Flask依赖

# 图像处理
opencv-python==4.10.0.84  # cv2
pillow==8.0.0  # PIL.Image

# 深度学习框架
torch==1.10.1+cu113  # 含CUDA支持（代码用torch、nn）
torchvision==0.11.2+cu113  # 图像处理转换（transforms）

# 数据处理
numpy==1.24.3  # 数组运算（cv2、torch均依赖）
pandas==1.3.4  # Excel处理（pd.read_excel）
openpyxl==3.1.2  # pandas读取Excel的依赖

# OpenAI接口
openai==1.98.0  # 调用OpenAI/DeepSeek接口

# 文件上传与编码
python-multipart==0.0.6  # Flask处理文件上传
base64==1.0.0  # 图像编码（系统库，通常无需安装，此处标记）